New instance system: Functional. Save instance_destroy().
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@263 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL/GScolors.h
@@ -46,8 +46,8 @@ enum {
   c_yellow  = 0x00FFFF
 };
 
-int draw_clear(double color);
-int draw_clear_alpha(double color, double alpha);
+int draw_clear(int color);
+int draw_clear_alpha(int color, float alpha);
 
 int merge_color(int col1, int col2, double amount);
